fullRoomNameを切り分け
types/RoomType.tsにRoomTypeの型定義を追加
utils/socket/SocketRoom.tsにgenerateFullRoomNameを作成
chat.gatewayのfullRoomNameをgenerateFullRoomNameに置き換え
