+++ backend/src/types/RoomType.ts
@@ -1,0 +1,1 @@
+export type RoomType = 'ChatRoom' | 'User' | 'Global';
